initial commit using meson for build flow
very rough draft, there be dragons.

The `tools/` directory Makefiles have been translated. In some instances there may
be compiler flags that were missed. Only tested on macos.

No make options outside of what meson provides as built-in were included either.
For example, in z64compress, the "native opt" option is not here and would need
to be included to pass `-march` to gcc/clang.

None of the build targets have the `install:` option. Because, for now I don't
have a good understanding for what should just be left as a build product or
what would be useful in an `install` location during decomp.

Unsure about next steps with some of the fixed paths that may need to be changed
or adjusted for asset extraction, running the rest of the build scripts, etc.
The approach was to just see what broke, why, and adjust/determine a strategy
based on what scripts expected what.